+++ providers/openai/openai.go
@@ -46,13 +46,18 @@ func New(opts ...config.Option) (*Provider, error) {
 // capabilities returns the capabilities for the OpenAI provider.
 func capabilities() providers.Capabilities {
 	return providers.Capabilities{
+		AsyncGeneration:     false,
 		Completion:          true,
+		CompletionAudio:     false,
 		CompletionImage:     true,
 		CompletionPDF:       false,
 		CompletionReasoning: true,
 		CompletionStreaming: true,
 		CompletionTools:     true,
+		CompletionVideo:     false,
 		Embedding:           true,
 		ListModels:          true,
+		STT:                 false,
+		TTS:                 false,
 	}
 }